[Fix] move file interface into core component
