+++ src/Monofony/Component/Core/Model/User/AdminAvatarInterface.php
@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Monofony\Component\Core\Model\User;
 
-use App\Entity\Media\FileInterface;
-
 interface AdminAvatarInterface extends FileInterface
 {
 }
